+++ mrv2/lib/mrvGL/mrvTimelineViewport.h
@@ -60,6 +60,8 @@ namespace mrv
         kRotate,
     };
 
+    enum MissingFrameType { kBlackFrame, kRepeatFrame, kScratchedFrame };
+
     class TimelinePlayer;
 
     class TimelineViewport : public Fl_SuperClass
@@ -229,6 +231,9 @@ namespace mrv
 
         //! Set the Annotation previous ghost frames.
         void setGhostNext(int);
+
+        //! Set the missing frame type.
+        void setMissingFrameType(const MissingFrameType);
 
         // Callbacks
         int acceptMultilineInput() noexcept;